requirements.txt
aiogram==2.13
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.9
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
typing_extensions==4.6.3
yarl==1.9.2
